+++ src/components/Dashboard/components/NutritionInfo/MealTypeDropDown.js
@@ -1,9 +1,6 @@
 import moment from "moment";
 import React, { useState } from "react";
 import { Dropdown } from "react-bootstrap";
-import DropdownButton from "react-bootstrap/DropdownButton";
-import FormControl from "react-bootstrap/FormControl";
-import InputGroup from "react-bootstrap/InputGroup";
 import { useDispatch, useSelector } from "react-redux";
 import { ReactComponent as LikeIcon } from "../../../../assets/LikeIcon.svg";
 import { addFoodToJournal } from "../../../../state/slices/userinfo";
